Checagem de tipos em chamadas de função melhorada
Alterado códigos .java para tornar compatível com os recursos
atualmente implementados no compilador

+++ java_src/TesteRecursao1.java
@@ -14,7 +14,8 @@ public class TesteRecursao1 {
        }
 	
 	public static void main(String[] args) {
-		int n ;
+		long n;
+		//int n ;
 		n = 10;
 		int a;
 		a = fibo(n-1);
