+++ java_src/TesteRecursao6.java
@@ -10,7 +10,8 @@ public class TesteRecursao6 {
 	}
 
 	public static void main(String[] args) {
-		int m;
+		//int m;
+		long m;
 		long a;
 		m = 25;
 		a = x(m);
